Retain 12 more GBC objects in full colour
Use per-8x8 CGB palette attributes, with a whole-object fallback whenever any colour is quantized so repeated sprite colours stay consistent.

Corpus: 95/125 -> 107/125 full-colour objects; cross-layer mappings 316 -> 228; inconsistent colours remain 0. Composition improves 16.8%-55.1% and initial rendering 18.3%-26.8%, at a cost of 84 bytes static WRAM.

+++ firmware/gbc/source/benchmark.h
@@ -23,7 +23,7 @@ typedef struct perf_interaction {
 
 void perfTimerStart(void);
 uint32_t perfTimerStop(void);
-uint8_t composeTile(uint32_t objects);
+uint16_t composeTile(uint32_t objects);
 void renderBoard(void);
 
 bool perfLoadFirstBoard(void) BANKED;
